Expose ydb_table CustomizeDiff in the public SDK package.
Move plan-time column validation behind sdk/terraform/table.CustomizeDiff so external consumers can reuse it without importing internal packages.
